requirements.txt
aiogram>=3.27,<4
APScheduler>=3.11,<4
python-dotenv>=1.0,<2
openai>=2.11,<3
Pillow>=10,<12
